Finish host.go

+++ host.go
@@ -14,8 +14,7 @@ import (
 )
 
 type submission_s struct{
-    Sid int
-    Uid int
+    Sid string 
     Submit_time string
     Code string
     Lang string
@@ -26,7 +25,7 @@ type submission_s struct{
 
 func httpGet() submission_s {
     var content submission_s
-    resp, err := http.Get("http://localhost:8080/api/getcode")
+    resp, err := http.Get("http://www.onionwyl.cn/api/getcode")
     if err != nil {
         // handle error
     }
@@ -36,15 +35,19 @@ func httpGet() submission_s {
     if err != nil {
         // handle error
     }
+    //fmt.Printf("%s\n", body)
     err = json.Unmarshal(body, &content)
-    //fmt.Println(content.Code);
+    //fmt.Println(content.Sid);
     return content
 }
 
 func main() {
     for{
         submission := httpGet()
-        fmt.Println(submission.Code)
+        if submission.Sid == ""{
+            continue
+        }
+        fmt.Println(submission)
         os.Remove("tmp.c")
         os.Remove("tmp.out")
         os.Remove("err")
@@ -56,7 +59,7 @@ func main() {
         cmd.Stdout = &out
         err := cmd.Run()
         if err != nil {
-            log.Fatal(err)
+           // log.Fatal(err)
         }
         cmd = exec.Command("/bin/sh", "-c", "./tmp.out")
         cmd.Stdin = strings.NewReader(submission.Input)
@@ -67,6 +70,18 @@ func main() {
         defer fin.Close()
         fd, _ := ioutil.ReadAll(fin)
         fmt.Printf("%s", fd)
+        submission.Output = out.String()
+        submission.Err_info = string(fd)
         time.Sleep(time.Second)
+        j, _ := json.Marshal(submission)
+        body := bytes.NewBuffer([]byte(j))
+        resp, err := http.Post("http://www.onionwyl.cn/api/putresult", "application/json;charset=utf-8", body)
+        if err != nil {
+            log.Fatal(err)
+            return
+        }
+       // result, _ := ioutil.ReadAll(resp.Body)
+        defer resp.Body.Close()
+       // fmt.Printf("%s", result)
     }
 }
